+++ 2023-10-otus-spring-solovyov/pom.xml
@@ -8,7 +8,7 @@
 		<version>3.1.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
+	<groupId>com.otus</groupId>
 	<artifactId>homework_2023_10_otus_spring_solovyov</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>2023-10-otus-spring-solovyov</name>
