+++ 2023-10-otus-spring-solovyov/src/main/java/com/otus/interfaces/QuestionManager.java
@@ -1,6 +1,6 @@
-package com.example.interfaces;
+package com.otus.interfaces;
 
-import com.example.model.Question;
+import com.otus.model.Question;
 
 import java.util.List;
