+++ 2023-10-otus-spring-solovyov/src/main/java/com/otus/interfaces/ReadFileContent.java
@@ -1,4 +1,4 @@
-package com.example.interfaces;
+package com.otus.interfaces;
 
 import java.util.List;
