+++ 2023-10-otus-spring-solovyov/src/main/java/com/otus/interfaces/ReadFileService.java
@@ -1,4 +1,4 @@
-package com.example.interfaces;
+package com.otus.interfaces;
 
 import java.io.File;
