+++ 2023-10-otus-spring-solovyov/src/main/java/com/otus/model/Question.java
@@ -1,4 +1,4 @@
-package com.example.model;
+package com.otus.model;
 
 public record Question(int id, String text) {
     @Override
