+++ 2023-10-otus-spring-solovyov/src/test/java/com/otus/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.otus;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
